requirements.txt
dj-database-url==0.5.0
Django==2.1.7
gunicorn==19.9.0
passlib==1.7.1
Pillow==5.4.1
pytz==2018.9
whitenoise==4.1.2
psycopg2==2.7.2
